% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.BA-package.R
\docType{package}
\name{sim.BA-package}
\alias{sim.BA}
\alias{sim.BA-package}
\title{sim.BA: Simulation-Based Bias Analysis for Observational Studies}
\description{
Allows user to conduct a simulation based quantitative bias analysis using covariate structures generated with individual-level data to characterize the bias arising from unmeasured confounding. Users can specify their desired data generating mechanisms to simulate data and quantitatively summarize findings in an end-to-end application using this package.
}
\author{
\strong{Maintainer}: Rishi Desai \email{rdesai@bwh.harvard.edu} (\href{https://orcid.org/0000-0003-0299-7273}{ORCID})

Authors:
\itemize{
  \item Noah Greifer \email{ngreifer@iq.harvard.edu} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID})
}

}
\keyword{internal}
