% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBA-package.R
\docType{package}
\name{simBA-package}
\alias{simBA-package}
\title{simBA: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Rishi Desai \email{rdesai@bwh.harvard.edu} (\href{https://orcid.org/0000-0003-0299-7273}{ORCID})

Authors:
\itemize{
  \item Noah Greifer \email{ngreifer@iq.harvard.edu} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID})
}

}
\keyword{internal}
